services/accounts/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==3.2.23
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.4
MarkupSafe==2.1.3
norminette==3.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
speedtest-cli==2.1.3
sqlparse==0.4.4
typing_extensions==4.1.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
wrapt==1.16.0
